requirements.txt
Jinja2
libvirt-python